Extract PageObject pattern implementation into a separate folder
