+++ src/__tests__/VisibilityToggle.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import "jest-styled-components";
-import AppPage from "./AppPage";
+import AppPage from "../testUtils/AppPageObject";
 
 jest.mock("../tree/sampleTrees", () => ({
   sampleNodes: jest.requireActual("../tree/sampleTrees").twoNestedNodes
@@ -8,7 +8,6 @@ jest.mock("../tree/sampleTrees", () => ({
 
 describe("Having a simple Tree", () => {
   const app = new AppPage();
-  afterEach(app.cleanup);
 
   it("should render two nodes", () => {
     expect(app.getAllNodes()).toHaveLength(2);
